Must specify a tag when copying to beta repo.

+++ copy-beta.sh
@@ -10,12 +10,24 @@ cd $DIR
 DEST=$1
 if [ "$DEST" = "" ]; then
   echo "No destination specified. Need /path/to/client-beta"
-  exit
+  exit 1
+fi
+
+TAG=$2
+if [ "$TAG" = "" ]; then
+  echo "No tag specified"
+  exit 1
+fi
+
+# Check dest exists
+if [ ! -d "$DEST" ]; then
+  echo "%DEST doesn't exist"
+  exit 2
 fi
 
 # Archive the current client repo
 rm -rf $DEST/client.tar
-git archive --format tar HEAD > $DEST/client.tar
+git archive --format tar $TAG > $DEST/client.tar
 
 # Move to destination dir
 cd $DEST
@@ -27,6 +39,20 @@ for i in "${dirs[@]}"; do
   rm -rf $i
 done
 
+echo "Unpacking git archive"
 tar xpf client.tar ${dirs[@]}
-
 rm client.tar
+
+
+echo "Now you should add, commit, tag and push the changes."
+echo "
+
+    cd $DEST
+    git add .
+    git commit -m \"Importing $TAG\"
+    git tag -a $TAG -m $TAG
+
+    git push
+    git push --tags
+
+"
